Create main
Mind that the elements in X and Y need to be processed column by column.
If the weight initialization is too small, it will cause the input of the neuron to be too small. As the number of layers increases, the signal will disappear. If the parameter initialization is too large, the input state will be too large. For the activation function, the value of the activation function will become saturated, resulting in the problem of the disappearance of the gradient.